mise a jour DateProd et initialisation des peremp

+++ src/abstraction/eq5Transformateur3/DateProd.java
@@ -3,9 +3,9 @@ package abstraction.eq5Transformateur3;
 // julien
 public class DateProd<Produit>{
 	private Produit produit; 
-	private int date; // la date est en fait le step auquel le produit sera perdu 
+	private double date; // la date est en fait le step auquel le produit sera perdu 
 	
-	public DateProd(int date,Produit produit) {
+	public DateProd(double date,Produit produit) {
 		this.produit=produit;
 		this.date=date;
 	}
@@ -13,7 +13,7 @@ public class DateProd<Produit>{
 	public Produit getProduit() {
 		return this.produit;
 	}
-	public int getDate() {
+	public double getDate() {
 		return this.date;
 	}
